+++ BolivianDaily.Domain/Repositories/ICategoryRepository.cs
@@ -4,5 +4,5 @@ using BolivianDaily.Domain.Entities;
 
 public interface ICategoryRepository
 {
-    Task<Category?> GetByIdAsync(int id, CancellationToken cancellationToken = default);
+    Task<Category?> GetByIdAsync(long id, CancellationToken cancellationToken = default);
 }
