+++ BolivianDaily.Domain/Repositories/ISourceRepository.cs
@@ -4,6 +4,6 @@ using BolivianDaily.Domain.Entities;
 
 public interface ISourceRepository
 {
-    Task<Source?> GetByIdAsync(int id, CancellationToken cancellationToken = default);
+    Task<Source?> GetByIdAsync(long id, CancellationToken cancellationToken = default);
     Task<IEnumerable<Source>> GetAllActiveAsync(CancellationToken cancellationToken = default);
 }
